pyproject.toml
[tool.poetry]
name = "Off-Policy Evaluation for Recommendations from Missing-Not-At-Random Rewards"
version = "0.1.0"
description = ""
authors = ["tatsuki takahashi"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
obp = "^0.5.5"
numpy = "^1.22.5"
scikit-learn = "1.1.3"
scipy = "1.9.3"
pandas = "^1.3.2"
seaborn = "^0.11.2"
matplotlib = "3.7.3"
hydra-core = "^1.3.2"
tqdm = "^4.66.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = 'py39'
line-length = 120

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
    "C90", # mccabe
]
ignore = ["E203", "E266"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
case-sensitive = true
force-single-line = true
force-sort-within-sections = true
lines-after-imports = 2
known-local-folder = [
    "dataset", 
    "ope", 
    "policy", 
    "utils",
    "db"
]
